solves #95 by testing whether the table exists before creation attempt (#98)

+++ database/migrations/2015_05_01_111209_hmt_tenants_table.php
@@ -3,6 +3,7 @@
 use Hyn\MultiTenant\Tenant\DatabaseConnection;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
 
 class HmtTenantsTable extends Migration
 {
@@ -13,24 +14,29 @@ class HmtTenantsTable extends Migration
      */
     public function up()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->create('tenants', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->string('customer_no')->nullable();
-            $table->string('name');
-            $table->string('email');
-            $table->boolean('administrator')->default(false);
+        if (!Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('tenants')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->create(
+                'tenants',
+                function (Blueprint $table) {
+                    $table->bigIncrements('id');
+                    $table->string('customer_no')->nullable();
+                    $table->string('name');
+                    $table->string('email');
+                    $table->boolean('administrator')->default(false);
 
-            $table->bigInteger('reseller_id')->unsigned()->nullable();
-            $table->bigInteger('referer_id')->unsigned()->nullable();
+                    $table->bigInteger('reseller_id')->unsigned()->nullable();
+                    $table->bigInteger('referer_id')->unsigned()->nullable();
 
-            $table->timestamps();
-            $table->softDeletes();
+                    $table->timestamps();
+                    $table->softDeletes();
 
-            $table->index(['customer_no', 'name']);
+                    $table->index(['customer_no', 'name']);
 
-            $table->foreign('reseller_id')->references('id')->on('tenants')->onDelete('set null');
-            $table->foreign('referer_id')->references('id')->on('tenants')->onDelete('set null');
-        });
+                    $table->foreign('reseller_id')->references('id')->on('tenants')->onDelete('set null');
+                    $table->foreign('referer_id')->references('id')->on('tenants')->onDelete('set null');
+                }
+            );
+        }
     }
 
     /**
@@ -40,6 +46,8 @@ class HmtTenantsTable extends Migration
      */
     public function down()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('tenants');
+        if (Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('tenants')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('tenants');
+        }
     }
 }
