+++ database/migrations/2015_05_01_111210_hmt_websites_table.php
@@ -1,8 +1,8 @@
 <?php
 
+use Hyn\MultiTenant\Tenant\DatabaseConnection;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
-use Hyn\MultiTenant\Tenant\DatabaseConnection;
 
 class HmtWebsitesTable extends Migration
 {
@@ -13,18 +13,23 @@ class HmtWebsitesTable extends Migration
      */
     public function up()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->create('websites', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('tenant_id')->unsigned();
-            $table->string('identifier');
-            $table->timestamps();
-            $table->softDeletes();
+        if (!Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('websites')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->create(
+                'websites',
+                function (Blueprint $table) {
+                    $table->bigIncrements('id');
+                    $table->bigInteger('tenant_id')->unsigned();
+                    $table->string('identifier');
+                    $table->timestamps();
+                    $table->softDeletes();
 
-            $table->foreign('tenant_id')->references('id')->on('tenants')->onDelete('cascade');
+                    $table->foreign('tenant_id')->references('id')->on('tenants')->onDelete('cascade');
 
-            $table->index('identifier');
-            $table->unique('identifier');
-        });
+                    $table->index('identifier');
+                    $table->unique('identifier');
+                }
+            );
+        }
     }
 
     /**
@@ -34,6 +39,8 @@ class HmtWebsitesTable extends Migration
      */
     public function down()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('websites');
+        if (Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('websites')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('websites');
+        }
     }
 }
