+++ database/migrations/2015_05_01_111211_hmt_hostnames_table.php
@@ -1,8 +1,8 @@
 <?php
 
+use Hyn\MultiTenant\Tenant\DatabaseConnection;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
-use Hyn\MultiTenant\Tenant\DatabaseConnection;
 
 class HmtHostnamesTable extends Migration
 {
@@ -13,36 +13,41 @@ class HmtHostnamesTable extends Migration
      */
     public function up()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->create('hostnames', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            // tenant owner
-            $table->bigInteger('tenant_id')->unsigned();
-            // hostname
-            $table->string('hostname');
-            // related to website x
-            $table->bigInteger('website_id')->unsigned()->nullable();
-            // related to certificate
-            $table->bigInteger('ssl_certificate_id')->unsigned()->nullable();
-            // subdomain of another hostname
-            $table->bigInteger('sub_of')->unsigned()->nullable();
-            // redirect to a different hostname
-            $table->bigInteger('redirect_to')->unsigned()->nullable();
-            // redirect standard to https if certificate available
-            $table->boolean('prefer_https')->default(false);
+        if (!Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('hostnames')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->create(
+                'hostnames',
+                function (Blueprint $table) {
+                    $table->bigIncrements('id');
+                    // tenant owner
+                    $table->bigInteger('tenant_id')->unsigned();
+                    // hostname
+                    $table->string('hostname');
+                    // related to website x
+                    $table->bigInteger('website_id')->unsigned()->nullable();
+                    // related to certificate
+                    $table->bigInteger('ssl_certificate_id')->unsigned()->nullable();
+                    // subdomain of another hostname
+                    $table->bigInteger('sub_of')->unsigned()->nullable();
+                    // redirect to a different hostname
+                    $table->bigInteger('redirect_to')->unsigned()->nullable();
+                    // redirect standard to https if certificate available
+                    $table->boolean('prefer_https')->default(false);
 
-            // timestaps
-            $table->timestamps();
-            $table->softDeletes();
+                    // timestaps
+                    $table->timestamps();
+                    $table->softDeletes();
 
-            // relations
-            $table->foreign('redirect_to')->references('id')->on('hostnames')->onDelete('set null');
-            $table->foreign('sub_of')->references('id')->on('hostnames')->onDelete('cascade');
-            $table->foreign('website_id')->references('id')->on('websites')->onDelete('set null');
-            $table->foreign('tenant_id')->references('id')->on('tenants')->onDelete('cascade');
+                    // relations
+                    $table->foreign('redirect_to')->references('id')->on('hostnames')->onDelete('set null');
+                    $table->foreign('sub_of')->references('id')->on('hostnames')->onDelete('cascade');
+                    $table->foreign('website_id')->references('id')->on('websites')->onDelete('set null');
+                    $table->foreign('tenant_id')->references('id')->on('tenants')->onDelete('cascade');
 
-            // index
-            $table->index('hostname');
-        });
+                    // index
+                    $table->index('hostname');
+                }
+            );
+        }
     }
 
     /**
@@ -52,6 +57,8 @@ class HmtHostnamesTable extends Migration
      */
     public function down()
     {
-        Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('hostnames');
+        if (Schema::connection(DatabaseConnection::systemConnectionName())->hasTable('hostnames')) {
+            Schema::connection(DatabaseConnection::systemConnectionName())->dropIfExists('hostnames');
+        }
     }
 }
